Move LabelWithTooltip to its own module
