+++ webapp/components/form/LabelWithTooltip/index.js
@@ -1,0 +1,1 @@
+export { LabelWithTooltip } from './LabelWithTooltip'
