+++ webapp/components/survey/SurveyForm/nodeDefs/components/nodeDefTableCellHeader.js
@@ -1,49 +1,15 @@
 import './nodeDefTableCellHeader.scss'
 
-import React, { useEffect, useRef, useState } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
 
 import * as NodeDef from '@core/survey/nodeDef'
 
+import { LabelWithTooltip } from '@webapp/components/form/LabelWithTooltip'
 import { useI18n } from '@webapp/store/system'
 
 import * as NodeDefUiProps from '../nodeDefUIProps'
 import NodeDefIconKey from './NodeDefIconKey'
-
-const LabelWithTooltip = (props) => {
-  const { label, style, children } = props
-
-  const labelRef = useRef(null)
-
-  // detect when ellipsis is active and show a tooltip in that case
-  const [ellipsed, setEllipsed] = useState(false)
-  const tooltip = ellipsed ? label : null
-
-  useEffect(() => {
-    const labelEl = labelRef?.current
-    if (labelEl) {
-      setEllipsed(labelEl.offsetHeight < labelEl.scrollHeight || labelEl.offsetWidth < labelEl.scrollWidth)
-    }
-  }, [label])
-
-  return (
-    <div className="label ellipsis" style={style} title={tooltip} ref={labelRef}>
-      {children}
-      {label}
-    </div>
-  )
-}
-
-LabelWithTooltip.propTypes = {
-  label: PropTypes.string.isRequired,
-  style: PropTypes.object,
-  children: PropTypes.node,
-}
-
-LabelWithTooltip.defaultProps = {
-  style: {},
-  children: null,
-}
 
 const NodeDefTableCellHeader = (props) => {
   const { label, nodeDef } = props
